Implemented some more complex AI for humans regarding their basic needs
Humans have an exhaustion, hunger, and NeedToUseToilet level that
increases over time. When these levels start rising to 5 or above, the
humans will automatically seek out the appropriate action to lower these
stats back down to 0. On the map, one human starts with a
NeedToUseToilet level of 5, while the other humans start with an
exhaustion level of 5. The exhausted humans decide to use beds and the
remaining human goes to the toilet. Once finished they then proceed to
move idly again unless/until another of these stats goes back up to 5.
The AI is able to tell whether a bed or a toilet is currently in use or
free. The AI will head to the nearest free bed/toilet.

I have yet to implement an appropriate action for hunger level to be
decreased so right now it just keeps going up and the AI doesn't really
know what to do when hunger level is high.