Names and Health
Added names and health UI above humans. Haven't added a way to randomly
generate names yet and currently there are no factors that lower health,
but both the name and health UI are completely functional and will
respond to health and name changes on humans.

Added functionality for occupiable actors such as toilets and beds to
teleport players and rotate them properly when they use the station.
Added a basic sleep animation which shows not just the animation change
but how the human is moved onto the correct position on the bed and then
moved back out when the human is done sleeping.

Laid some groundwork for weapons.